apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hero-ai-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlmodel>=0.0.19",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.7.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "google-generativeai>=0.7.0",
    "tenacity>=8.3.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
    "Pillow>=10.0.0",
    "pydantic-settings>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
